// code/frontend/src/app/models/dto/ProdutoCadastroDTO.ts
export interface ProdutoCadastroDTO {
  nome: string;
  precoCusto: number;
  margemLucro: number;
  precoFinalVenda: number;
  idCategoria: number;
  quantidadeDisponivelEstoque: number;
  quantidadeMinimaEstoque: number;
  idUnidadeMedida: number;
  quantidadePorMedida: number;
  dataVencimento?: Date;
  codigoBarras?: string;
}